backend/requirements.txt
cycler==0.10.0
Django==2.2.4
djangorestframework==3.10.2
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
